[feat] Add Shelly-1L configuration example (#54)
* [feat] Add Shelly-1L configuration example

* [fix] Add required indentation around headers

* [fix] Add blank line around code block and remove trailing space

* [rev] Add spacing around sentences

* [fix] Remove another trailing space...

Is it not possible to have the linter auto-fix these small errors instead of reporting them? Not everyone will have a linter or a linter with the same config as this project expects. I also don't see how an extra space after a dot would do any harm in a markdown file.