Added notice to the baud_rate (#318)
I had the issue that was noted above in the 'notes', but I didnt read it an only copy and pasted the yaml file.  this adds some comments so you see the issue in the yaml file. I think all new DS02S 3 way wifi dimmer's require 115200.  Maybe we should go ahead and set the default to 115200 instead of making it a note.