Corrected which Tuya Module comes from factory (#602)
* Corrected chip type to RTL8710

I had replaced my modules with ESP8266 and *assumed* what I replaced was Beken modules but they are, rather, Realtek RTL8710BN modules.  

I have corrected this, and added separate YAML for using the WR4 or replacing it with an ESP8266.

* Add files via upload

* Added images of the fan and the controller board

* Removed blank lines

I have a habit of leaving multiple blank lines in my Markdown.