Update index.md to address issues with GPIO16 & relay1
From my own experiences with this board as well as the "more detailed manual" and this guide here: https://templates.blakadder.com/ESP12F_Relay_X4.html

GPIO16 is used for RXD on this board so it will always be briefly powered on at boot. This may be an issue for someone's use case so it should be addressed in documentation. I have found the suggested use of GPIO15 on the tasmota html page to allow relay1 to function normally.