Update Sonoff-POW-Elite-20a/index.md
Removed the hall sensor, as its deprecated. Changed to esp-idf framework because of limited flash memory, and moved to esp internal_temperature platform. Also added ignore for the strapping pins as this is a prebuilt unit. Also changed the deprecated publish_initial_state to its equivalent trigger_on_initial_state