Update Lanbon L9 configuration (#1227)
The L9-HS has been updated (my PCB shows 20250225) and the display will not work unless the reset pin is added to the display config : https://i.postimg.cc/t45gvdB2/l9v2.jpg
I do not know if this impacts the previous version of this device (physical case also slightly different) as I do not have one to test.
Adding "reset_pin: GPIO4" enabled the device to boot and work as expected. Without this it would not boot.