Update index.md (#1342)
My relay shows a phantom voltage of about 75V when it is turned off and there is no load.  So I think the template relay state on boot needs to be set with a threshold of 90 or so (which is a guess, but it works for me) if no current is detected, while using the lower voltage threshold of 10 if a current is detected.  The latter condition accounts for the possibility of extremely low mains voltages as long as some load is detected.  

Also, the dynamic template relay state doesn't need to be optimistic based on this; I think it should be set based on the same lambda that determines its state on startup. 

(In my case, using the previous strict limit of 10V to determine if the relay is on at startup, the relay was always detected as on, and hence the code would always actually turn the relay on at startup.)